Adjusted SSIM and PSNR metrics in DeblurGAN

+++ cnn-deblur/models/deblur_gan.py
@@ -285,10 +285,10 @@ class DeblurGan(Model):
 
         # Compute metrics
         ssim_metric = tf.image.ssim(sharp_batch,
-                                    tf.cast(generated_batch, dtype='bfloat16'),
+                                    generated_batch,
                                     max_val=2.)
         psnr_metric = tf.image.psnr(sharp_batch,
-                                    tf.cast(generated_batch, dtype='bfloat16'),
+                                    generated_batch,
                                     max_val=2.)
 
         return {"val_d_loss": d_loss,
